requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
moviepy>=1.0.3
srt>=3.5.3
imageio-ffmpeg>=0.4.9
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
